chore(HttpClient)!: keep User-Agent on headers
If User-Agent is given via headers, but is not given via options, keep
the headers User-Agent instead of replacing it with restify-clients UA.
This change makes User-Agent handling more aligned with how
restify-clients deals with other headers which can be overriten by
options.

The reasoning behind this is in cases where restify-clients is acting
like a forward-proxy of sorts. If User-Agent is present in headers
there's probably a request forwarding happening.